Add plugin configuration
The "gerrit.config" file in this patchset contains a minimal
configuration template. Once the placeholders are replaced with actual
values, its content can be appended to the gerrit.config file located at
"<path_to_gerrit_instance>/etc/gerrit.config", for global configuration,
or to the project.config file of the repository, for project-specific
configuration.

Jira-Id: IT-103
Change-Id: Ibfdfebcd1b05933069e7fa640accefe7ab9e879d
Signed-off-by: Patrizio <patrizio.gelosi@amarulasolutions.com>