Support ChatGPT response to inline comments
Added the capability to reply to multiple inline comments, minimizing
the payload sent to ChatGPT. The prompt sent to ChatGPT is adapted so
that each comment is converted into a tagged request. This process
includes incorporating the selected code snippet in each request to
enhance the insight quality of ChatGPT. Subsequently, the related
responses from ChatGPT are extracted and submitted back to Gerrit, with
each response linked to its corresponding request.
NOTES:
- Due to the absence of a specific Gerrit REST endpoint to fetch
comments from a processed 'comment-added' event, the general GET
comments endpoint (returning all comments in a patchset) is used
instead. Comments are filtered to include only those made by the same
author as the event's author, within a certain time period prior to the
event.
- State retention issues occurring after subsequent mixed patchset
submissions and inline comment requests have been resolved using the
`initialize` method of the `GerritClientBase` class.
- The splitting of ChatGPT comments longer than a given number of
lines into multiple comments is suppressed.

Jira-Id: IT-103
Change-Id: I0d78e47fc5e3aad1ef3e6d449dcd1bded28c2abc
Signed-off-by: Patrizio <patrizio.gelosi@amarulasolutions.com>