Add DisabledAuthors/EnabledAuthors config options
Added the `enabledAuthors` option, allowing for the specification of
users authorized to have their Patchsets and comments reviewed. By
default, all users are enabled, but this setting can be adjusted to
restrict reviews to specific usernames.
Additionally, the disabledAuthors option has been added, serving as a
counterpart to `enabledAuthors` with inverse functionality.

Jira-Id: IT-103
Change-Id: Ie0dcda50ea5f46e6000de1569a2c2fbdf76f4466
Signed-off-by: Patrizio <patrizio.gelosi@amarulasolutions.com>