Add Disabled/Enabled-Groups config options
Implemented the `enabledGroups` option, enabling the specification of
user groups eligible for Patchset and comment reviews. This option is
set to include all groups by default, but can be tailored to limit
reviews to selected groups. Alongside, the `disabledGroups` option has
been introduced as a complementary feature to `enabledGroups`, providing
reverse functionality.

Jira-Id: IT-103
Change-Id: Ib92ff4e4fa6ad0fc03e5dc3d70a03108282853ee
Signed-off-by: Patrizio <patrizio.gelosi@amarulasolutions.com>