Improve multi-message chat interaction
Addressed occasional issues with out-of-context responses in
multi-message chats with ChatGPT. Earlier, responses could be off-target
because they did not include the entire history of previous replies.
With this update, ChatGPT has access to the full conversation history
related to each message where it is tagged, leading to more relevant and
accurate responses.

Jira-Id: IT-103
Change-Id: Ie36c77a6f6e4e89c66c33672e5b55c2680903358
Signed-off-by: Patrizio <patrizio.gelosi@amarulasolutions.com>