Introduce `/directive` command support
Enabled the use of the `/directive` command in comments directed at
ChatGPT. When a comment in a Change Set includes this command along with
a specific directive, it forces ChatGPT to adhere to this directive in
subsequent reviews of the Change Set.
This enhancement addresses the previously lacking ability to directly
guide ChatGPT through comments. Specifically, it resolves issues where
ChatGPT's actions contradicted its own statements, such as agreeing to a
directive but then sending a review that disregards it.
An example of using the directive command to halt further commit message
reviews in the Change Set is:

@gpt /directive do not review the commit message

Jira-Id: IT-103
Change-Id: Icfabd76147903fd4abb10fc4f9303a6c77b1b1d3
Signed-off-by: Patrizio <patrizio.gelosi@amarulasolutions.com>