Support full-cycle stateful review requests
Full-cycle stateful patch-set review requests are now supported through
the implementation of the ChatGPT Run layer. Once the initial message is
committed to a new ChatGPT thread, a ChatGPT run is created and linked
to that thread.
This run is then polled at regular intervals (default: 1 second) until
completion. Subsequently, the steps of the run are retrieved along with
the ChatGPT replies, which are processed similarly to the stateless
mode.
Currently, the following stateless features are not yet supported in
stateful mode:
- Review of commit messages
- Inline code lookup
- Detection of repeated replies

Change-Id: I84f7087d838f0cec9bc55c8e724519db4f17df72
Signed-off-by: Patrizio <patrizio.gelosi@amarulasolutions.com>