Implement localization framework
A comprehensive localization framework is introduced to the project,
transitioning from the use of hardcoded language constants in the
Settings class to a dynamic, locale-specific ResourceBundle approach.
The changes include:

- Creation of the Localizer class that is now injected through Guice to
provide dynamic language support based on user locale. This approach
leverages the ResourceBundle mechanism for localization.
- Addition of ResourceBundle property files for English (default),
organized by specific namespaces to facilitate easy expansion and
maintenance.
- Refactoring throughout the application to replace direct usage of
language-specific constants from the Settings class with calls to the
Localizer class. This shift enhances maintainability and flexibility in
language support.

Change-Id: Ibe343fe6d2fee8cd799ab2196df22c6940335cc2
Signed-off-by: Patrizio <patrizio.gelosi@amarulasolutions.com>