Implement task-specific assistants
Different assistants are now specialized for various tasks within the
same thread. This commit introduces an assistant specialized in handling
reviews and another dedicated to managing requests directed at ChatGPT.
Additionally, this commit modifies the instantiation of the review
assistant, which previously occurred upon patchset creation. It is now
initialized during the run creation process, aligning with the setup for
the requests assistant.

Change-Id: I50aa362fa56dbb095f653eeedb4a6788befc52e9
Signed-off-by: Patrizio <patrizio.gelosi@amarulasolutions.com>